Update README.md
This mod is only for people who speedrun the game. Don't use this, if you want to play the game casually.

What does this mod do?
If you click on Restart level, it will teleport you to the first room and restart the game.

Why should I use this?
If you don't, you will always have to close and open the game, which can get annoying.

Is this allowed in runs?
So far, this is the only mod, that is allowed to be used in runs.

What is Assembly-CSharp.dll?
Assembly-CSharp.dll has all the scripts that were written in Unity in it.

I would want to make a mod for this game. How can I make one?
This video is not for Grey-Box Testing, but it should teach you the basics:
https://www.youtube.com/watch?v=gV16BPG550M
Note, that you have to know basics of C# to make mods.

How do I install it?
Go to Grey_Box_Testing_0_1_5 (the folder where the game is located) > Managed > Grey-Box Testing Data and replace Assembly-CSharp.dll with Assembly-CSharp.dll. If you did everything correctly, the game should read this DLL instead of the normal one.

In the file you downloaded from github, there is also another Assembly-CSharp - original.dll. This is the original DLL that you download with the game. In case you want to delete the mod, just rename Assembly-CSharp - original.dll to just Assembly-CSharp.dll and replace it in Grey_Box_Testing_0_1_5 > Managed > Grey-Box Testing_Data.